Refactored abstract validate and added abstract data type
For schema validations the majority of the logic in abstract validate
was needed. So code was refactored to better represent the name spaces.

Task #15
